feat: add first
This file contains a cpp script for building a bluetooth connection via SPP under the name "Wassersensor". It can measure the amount of water going through the sensor. Also there is a button implemented which needs to be pressed, when the direction of the waterflow is changed. Every second the ESP32 sends a JSON with the overall water volume.